feat: parsing module, bump version
