+++ docs/bigtree/dag/parsing.md
@@ -1,0 +1,5 @@
+---
+title: DAG Parsing
+---
+
+::: bigtree.dag.parsing
