+++ docs/bigtree/tree/parsing.md
@@ -1,0 +1,5 @@
+---
+title: Tree Parsing
+---
+
+::: bigtree.tree.parsing
