+++ tests/node/test_dagnode.py
@@ -1,6 +1,5 @@
 import copy
 import unittest
-from itertools import combinations
 
 import pandas as pd
 import pytest
@@ -617,102 +616,6 @@ class TestDAGNode(unittest.TestCase):
                     assert list(child.parents) == [
                         parent
                     ], f"Node {child} parent, expected {parent}, received {child.parents}"
-
-    def test_go_to(self):
-        self.a >> self.b
-        self.b >> self.c
-        self.b >> self.d
-        self.c >> self.e
-        self.c >> self.f
-        self.e >> self.f
-        self.f >> self.g
-        self.d >> self.e
-
-        expected_paths = [
-            [["a", "b"]],
-            [["a", "b", "c"]],
-            [["a", "b", "d"]],
-            [["a", "b", "c", "e"], ["a", "b", "d", "e"]],
-            [
-                ["a", "b", "c", "e", "f"],
-                ["a", "b", "c", "f"],
-                ["a", "b", "d", "e", "f"],
-            ],
-            [
-                ["a", "b", "c", "e", "f", "g"],
-                ["a", "b", "c", "f", "g"],
-                ["a", "b", "d", "e", "f", "g"],
-            ],
-            [["b", "c"]],
-            [["b", "d"]],
-            [["b", "c", "e"], ["b", "d", "e"]],
-            [["b", "c", "e", "f"], ["b", "c", "f"], ["b", "d", "e", "f"]],
-            [
-                ["b", "c", "e", "f", "g"],
-                ["b", "c", "f", "g"],
-                ["b", "d", "e", "f", "g"],
-            ],
-            None,
-            [["c", "e"]],
-            [["c", "e", "f"], ["c", "f"]],
-            [["c", "e", "f", "g"], ["c", "f", "g"]],
-            [["d", "e"]],
-            [["d", "e", "f"]],
-            [["d", "e", "f", "g"]],
-            [["e", "f"]],
-            [["e", "f", "g"]],
-            [["f", "g"]],
-        ]
-        for node_pair, expected_path in zip(
-            combinations(self.nodes, 2), expected_paths
-        ):
-            if not expected_path:
-                with pytest.raises(exceptions.TreeError) as exc_info:
-                    node_pair[0].go_to(node_pair[1])
-                assert str(exc_info.value) == Constants.ERROR_NODE_GOTO.format(
-                    node=node_pair[1]
-                )
-            else:
-                actual_path = [
-                    [_node.node_name for _node in _path]
-                    for _path in node_pair[0].go_to(node_pair[1])
-                ]
-                assert (
-                    actual_path == expected_path
-                ), f"Wrong path for {node_pair}, expected {expected_path}, received {actual_path}"
-
-    def test_go_to_same_node(self):
-        self.a >> self.b
-        self.b >> self.c
-        self.b >> self.d
-        self.c >> self.e
-        self.c >> self.f
-        self.e >> self.f
-        self.f >> self.g
-        self.d >> self.e
-
-        for _node in self.nodes:
-            actual_path = [
-                [_node2.node_name for _node2 in _node1] for _node1 in _node.go_to(_node)
-            ]
-            expected_path = [[_node.node_name]]
-            assert (
-                actual_path == expected_path
-            ), f"Wrong path for {_node}, expected {expected_path}, received {actual_path}"
-
-    def test_go_to_type_error(self):
-        destination = 2
-        with pytest.raises(TypeError) as exc_info:
-            self.a.go_to(destination)
-        assert str(exc_info.value) == Constants.ERROR_NODE_GOTO_TYPE.format(
-            type="DAGNode", input_type=type(destination)
-        )
-
-    def test_go_to_different_tree_error(self):
-        a = dagnode.DAGNode("a")
-        with pytest.raises(exceptions.TreeError) as exc_info:
-            a.go_to(self.a)
-        assert str(exc_info.value) == Constants.ERROR_NODE_GOTO.format(node=self.a)
 
 
 def assert_dag_structure_root(dag):
